+++ stroom-core-client/src/main/java/stroom/cache/client/presenter/CacheNodeListPresenter.java
@@ -19,8 +19,7 @@ package stroom.cache.client.presenter;
 import stroom.cache.shared.CacheInfo;
 import stroom.cache.shared.CacheInfoResponse;
 import stroom.cache.shared.CacheResource;
-import stroom.cell.info.client.InfoColumn;
-import stroom.data.client.presenter.ColumnSizeConstants;
+import stroom.cell.info.client.ActionCell;
 import stroom.data.client.presenter.RestDataProvider;
 import stroom.data.grid.client.DataGridView;
 import stroom.data.grid.client.DataGridViewImpl;
@@ -30,37 +29,50 @@ import stroom.dispatch.client.RestFactory;
 import stroom.node.client.NodeManager;
 import stroom.svg.client.SvgPreset;
 import stroom.svg.client.SvgPresets;
+import stroom.util.client.DataGridUtil;
 import stroom.util.client.DelayedUpdate;
+import stroom.util.shared.ModelStringUtil;
 import stroom.util.shared.PageResponse;
-import stroom.widget.popup.client.event.ShowPopupEvent;
-import stroom.widget.popup.client.presenter.PopupPosition;
-import stroom.widget.popup.client.presenter.PopupView.PopupType;
 import stroom.widget.tooltip.client.presenter.TooltipPresenter;
-import stroom.widget.tooltip.client.presenter.TooltipUtil;
 
-import com.google.gwt.cell.client.ButtonCell;
+import com.google.gwt.cell.client.Cell.Context;
 import com.google.gwt.cell.client.TextCell;
 import com.google.gwt.core.client.GWT;
-import com.google.gwt.safehtml.shared.SafeHtml;
+import com.google.gwt.dom.client.Element;
+import com.google.gwt.dom.client.NativeEvent;
+import com.google.gwt.i18n.client.NumberFormat;
+import com.google.gwt.regexp.shared.RegExp;
 import com.google.gwt.user.cellview.client.Column;
 import com.google.gwt.view.client.Range;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 import com.gwtplatform.mvp.client.MyPresenterWidget;
 
+import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.function.BiConsumer;
 import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 public class CacheNodeListPresenter extends MyPresenterWidget<DataGridView<CacheInfo>> {
 
     private static final CacheResource CACHE_RESOURCE = GWT.create(CacheResource.class);
 
+    private static final int ICON_COL = 18;
     private static final int SMALL_COL = 90;
-    private static final int MEDIUM_COL = 150;
+    private static final int MEDIUM_COL = 130;
+    private static final RegExp CASE_CONVERSION_REGEX = RegExp.compile("([a-z])([A-Z])", "g");
+    protected static final String CACHE_INFO_KEY_HIT_COUNT = "HitCount";
+    protected static final String CACHE_INFO_KEY_MISS_COUNT = "MissCount";
+    protected static final String HIT_RATIO_KEY = "HitRatio";
 
     private final RestFactory restFactory;
     private final TooltipPresenter tooltipPresenter;
@@ -75,6 +87,8 @@ public class CacheNodeListPresenter extends MyPresenterWidget<DataGridView<Cache
 
     private Range range;
     private Consumer<CacheInfoResponse> dataConsumer;
+    private Set<String> cacheInfoKeys = new HashSet<>();
+    private List<Column<CacheInfo, ?>> columns = new ArrayList<>();
 
     @Inject
     public CacheNodeListPresenter(final EventBus eventBus,
@@ -86,97 +100,176 @@ public class CacheNodeListPresenter extends MyPresenterWidget<DataGridView<Cache
         this.tooltipPresenter = tooltipPresenter;
         this.nodeManager = nodeManager;
         this.delayedUpdate = new DelayedUpdate(this::update);
+    }
 
-        // Info.
-        addInfoColumn();
+    private void addColumns() {
+        for (final Column<CacheInfo, ?> column : columns) {
+            getView().removeColumn(column);
+        }
+        columns.clear();
+
+        // Clear.
+        addIconButtonColumn(
+                SvgPresets.of(SvgPresets.DELETE, "Clear cache", true),
+                (row, nativeEvent) -> {
+                    final Rest<Boolean> rest = restFactory.create();
+                    rest
+                            .onSuccess(result -> {
+                                dataProvider.refresh();
+                            })
+                            .call(CACHE_RESOURCE)
+                            .clear(row.getName(), row.getNodeName());
+                });
 
         // Node.
-        getView().addResizableColumn(new Column<CacheInfo, String>(new TextCell()) {
+        addColumn(new Column<CacheInfo, String>(new TextCell()) {
             @Override
             public String getValue(final CacheInfo row) {
                 return row.getNodeName();
             }
         }, "Node", MEDIUM_COL);
 
-        // Entries.
-        getView().addResizableColumn(new Column<CacheInfo, String>(new TextCell()) {
-            @Override
-            public String getValue(final CacheInfo row) {
-                return row.getMap().get("Entries");
-            }
-        }, "Entries", SMALL_COL);
+        final List<String> sortedCacheKeys = new ArrayList<>(cacheInfoKeys);
+        sortedCacheKeys.add(HIT_RATIO_KEY);
+        sortedCacheKeys.sort(Comparator.naturalOrder());
 
-        // Max Entries.
-        getView().addResizableColumn(new Column<CacheInfo, String>(new TextCell()) {
-            @Override
-            public String getValue(final CacheInfo row) {
-                return row.getMap().get("MaximumSize");
-            }
-        }, "Max Entries", SMALL_COL);
+        for (final String cacheInfoKey : sortedCacheKeys) {
+            final String name = convertUpperCamelToHuman(cacheInfoKey);
 
-        // Expiry.
-        getView().addResizableColumn(new Column<CacheInfo, String>(new TextCell()) {
-            @Override
-            public String getValue(final CacheInfo row) {
-                String expiry = row.getMap().get("ExpireAfterAccess");
-                if (expiry == null) {
-                    expiry = row.getMap().get("ExpireAfterWrite");
-                }
-                return expiry;
+            if (HIT_RATIO_KEY.equals(cacheInfoKey)
+                    && cacheInfoKeys.contains(CACHE_INFO_KEY_HIT_COUNT)
+                    && cacheInfoKeys.contains(CACHE_INFO_KEY_MISS_COUNT)) {
+                addStatColumn("Hit Ratio", -1, row ->
+                        getCacheHitRatio(row.getMap()));
+            } else {
+                addStatColumn(name, -1, row -> {
+                    String value = row.getMap().get(cacheInfoKey);
+                    return formatValue(cacheInfoKey, value);
+                });
             }
-        }, "Expiry", SMALL_COL);
+        }
 
-        // Clear.
-        final Column<CacheInfo, String> clearColumn = new Column<CacheInfo, String>(new ButtonCell()) {
-            @Override
-            public String getValue(final CacheInfo row) {
-                return "Clear";
-            }
-        };
-        clearColumn.setFieldUpdater((index, row, value) -> {
-            final Rest<Boolean> rest = restFactory.create();
-            rest.call(CACHE_RESOURCE).clear(row.getName(), row.getNodeName());
-        });
-        getView().addColumn(clearColumn, "</br>", 50);
-
-        getView().addEndColumn(new EndColumn<>());
+        final EndColumn<CacheInfo> endColumn = new EndColumn<>();
+        columns.add(endColumn);
+        getView().addEndColumn(endColumn);
     }
 
-    private void addInfoColumn() {
-        // Info column.
-        final InfoColumn<CacheInfo> infoColumn = new InfoColumn<CacheInfo>() {
-            @Override
-            public SvgPreset getValue(final CacheInfo object) {
-                return SvgPresets.INFO;
-            }
+    private void addIconButtonColumn(final SvgPreset svgPreset,
+                                     final BiConsumer<CacheInfo, NativeEvent> action) {
+        final ActionCell<CacheInfo> cell = new stroom.cell.info.client.ActionCell<CacheInfo>(
+                svgPreset, action);
+        final Column<CacheInfo, CacheInfo> col =
+                new Column<CacheInfo, CacheInfo>(cell) {
+                    @Override
+                    public CacheInfo getValue(final CacheInfo row) {
+                        return row;
+                    }
 
-            @Override
-            protected void showInfo(final CacheInfo row, final int x, final int y) {
-                final SafeHtml html = getInfoHtml(row);
-                tooltipPresenter.setHTML(html);
-                final PopupPosition popupPosition = new PopupPosition(x, y);
-                ShowPopupEvent.fire(CacheNodeListPresenter.this, tooltipPresenter, PopupType.POPUP,
-                        popupPosition, null);
-            }
-        };
-        getView().addColumn(infoColumn, "<br/>", ColumnSizeConstants.ICON_COL);
+                    @Override
+                    public void onBrowserEvent(final Context context,
+                                               final Element elem,
+                                               final CacheInfo rule,
+                                               final NativeEvent event) {
+                        super.onBrowserEvent(context, elem, rule, event);
+                    }
+                };
+        columns.add(col);
+        getView().addColumn(col, "", ICON_COL);
     }
 
-    private SafeHtml getInfoHtml(final CacheInfo cacheInfo) {
+    private void addStatColumn(final String name,
+                               final int width,
+                               final Function<CacheInfo, String> valueExtractor) {
 
-        return TooltipUtil.builder()
-                .addTwoColTable(tableBuilder -> {
-                    tableBuilder.addHeaderRow(cacheInfo.getNodeName());
-                    final Map<String, String> map = cacheInfo.getMap();
-                    map.keySet().stream()
-                            .sorted(Comparator.naturalOrder())
-                            .forEachOrdered(k -> {
-                                final String v = map.get(k);
-                                tableBuilder.addRow(k, v);
-                            });
-                    return tableBuilder.build();
-                })
+        final Column<CacheInfo, String> col = DataGridUtil.columnBuilder(
+                        valueExtractor,
+                        Function.identity(),
+                        TextCell::new)
+                .rightAligned()
                 .build();
+        columns.add(col);
+        final int newWidth = width == -1
+                ? determineColumnWidth(name)
+                : width;
+        getView().addResizableColumn(
+                col,
+                DataGridUtil.createRightAlignedHeader(name),
+                newWidth);
+    }
+
+    private String getCacheHitRatio(final Map<String, String> cacheInfo) {
+        final long hitCount = Long.parseLong(cacheInfo.get(CACHE_INFO_KEY_HIT_COUNT));
+        final long missCount = Long.parseLong(cacheInfo.get(CACHE_INFO_KEY_MISS_COUNT));
+        if (hitCount + missCount == 0) {
+            return "-";
+        } else {
+            final double ratio = ((double) hitCount) / (hitCount + missCount);
+
+            final BigDecimal scaledVal = new BigDecimal(ratio).setScale(3, RoundingMode.HALF_UP);
+            return NumberFormat.getDecimalFormat()
+                    .format(scaledVal);
+        }
+    }
+
+    private String formatValue(final String cacheKey, final String value) {
+        if (cacheKey.equals("TotalLoadTime")) {
+            return value;
+        } else if (cacheKey.startsWith("ExpireAfter")) {
+            return value;
+        } else {
+            try {
+                return ModelStringUtil.formatCsv(Long.parseLong(value));
+            } catch (NumberFormatException e) {
+                return value;
+            }
+        }
+    }
+
+    private String convertUpperCamelToHuman(final String str) {
+        return CASE_CONVERSION_REGEX.replace(str, "$1 $2");
+    }
+
+    private void addColumn(Column<CacheInfo, ?> column, String name, int width) {
+        columns.add(column);
+        final int newWidth = width == -1
+                ? determineColumnWidth(name)
+                : width;
+        getView().addResizableColumn(column, name, newWidth);
+    }
+
+    /**
+     * Crude method to size a column by its header
+     */
+    private int determineColumnWidth(final String text) {
+        // Not a fixed width font so some chars are narrower
+        final long thinCharsCount = text.chars()
+                .filter(chr ->
+                        chr == 'I'
+                                || chr == 'i'
+                                || chr == 'j'
+                                || chr == 'l'
+                                || chr == 'r'
+                                || chr == 't')
+                .count();
+        final long wideCharsCount = text.chars()
+                .filter(chr ->
+                        chr == 'M'
+                                || chr == 'W'
+                                || chr == 'm'
+                                || chr == 'w')
+                .count();
+        final long normalCharsCount = text.length() - thinCharsCount - wideCharsCount;
+        // Adjust for the narrow chars present
+        final double adjustedCharsCount = normalCharsCount
+                + ((double) thinCharsCount * 0.6)
+                + ((double) wideCharsCount * 1.3);
+        // Now scale the to pixels
+        final int colWidth = (int) (adjustedCharsCount * 8.5);
+//        GWT.log("text: " + text
+//                + " length: " + text.length()
+//                + " eff width: " + effectiveWidth
+//                + " col width: " + colWidth);
+        return Math.max(colWidth, SMALL_COL);
     }
 
     public void read(final String cacheName) {
@@ -207,11 +300,17 @@ public class CacheNodeListPresenter extends MyPresenterWidget<DataGridView<Cache
     private void fetchTasksForNodes(final Consumer<CacheInfoResponse> dataConsumer,
                                     final Consumer<Throwable> throwableConsumer,
                                     final List<String> nodeNames) {
+        cacheInfoKeys.clear();
         for (final String nodeName : nodeNames) {
             final Rest<CacheInfoResponse> rest = restFactory.create();
             rest
                     .onSuccess(response -> {
                         responseMap.put(nodeName, response.getValues());
+
+                        cacheInfoKeys.addAll(response.getValues()
+                                .stream()
+                                .flatMap(resp -> resp.getMap().keySet().stream())
+                                .collect(Collectors.toSet()));
                         delayedUpdate.update();
                     })
                     .onFailure(throwable -> {
@@ -224,6 +323,7 @@ public class CacheNodeListPresenter extends MyPresenterWidget<DataGridView<Cache
 
     private void update() {
         // Combine data from all nodes.
+        addColumns();
         final List<CacheInfo> list = new ArrayList<>();
         responseMap.values().forEach(list::addAll);
         list.sort(Comparator.comparing(CacheInfo::getName));